worker/requirements.txt
pika
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
